2. Mini-Projects/7. MedicalQA-RAG/requirements.txt
qdrant-client
openai
tiktoken
streamlit
python-dotenv
langchain
langchain_community
langchain-core 
langchain-openai 
chromadb